+++ pkg/oc/cli/admin/admin.go
@@ -13,32 +13,32 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	ktemplates "k8s.io/kubernetes/pkg/kubectl/util/templates"
 
+	"github.com/openshift/oc/pkg/cli/admin/cert"
+	"github.com/openshift/oc/pkg/cli/admin/createerrortemplate"
+	"github.com/openshift/oc/pkg/cli/admin/createlogintemplate"
+	"github.com/openshift/oc/pkg/cli/admin/createproviderselectiontemplate"
+	"github.com/openshift/oc/pkg/cli/admin/migrate"
+	migrateetcd "github.com/openshift/oc/pkg/cli/admin/migrate/etcd"
+	migratehpa "github.com/openshift/oc/pkg/cli/admin/migrate/legacyhpa"
+	migratestorage "github.com/openshift/oc/pkg/cli/admin/migrate/storage"
+	migratetemplateinstances "github.com/openshift/oc/pkg/cli/admin/migrate/templateinstances"
+	"github.com/openshift/oc/pkg/cli/admin/network"
+	"github.com/openshift/oc/pkg/cli/admin/node"
+	"github.com/openshift/oc/pkg/cli/admin/upgrade"
 	"github.com/openshift/oc/pkg/cli/options"
 	cmdutil "github.com/openshift/oc/pkg/helpers/cmd"
 	"github.com/openshift/origin/pkg/cmd/templates"
 	"github.com/openshift/origin/pkg/oc/cli/admin/buildchain"
-	"github.com/openshift/origin/pkg/oc/cli/admin/cert"
 	"github.com/openshift/origin/pkg/oc/cli/admin/createbootstrapprojecttemplate"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createerrortemplate"
 	"github.com/openshift/origin/pkg/oc/cli/admin/createkubeconfig"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createlogintemplate"
-	"github.com/openshift/origin/pkg/oc/cli/admin/createproviderselectiontemplate"
 	"github.com/openshift/origin/pkg/oc/cli/admin/groups"
-	"github.com/openshift/origin/pkg/oc/cli/admin/migrate"
-	migrateetcd "github.com/openshift/origin/pkg/oc/cli/admin/migrate/etcd"
 	migrateimages "github.com/openshift/origin/pkg/oc/cli/admin/migrate/images"
-	migratehpa "github.com/openshift/origin/pkg/oc/cli/admin/migrate/legacyhpa"
-	migratestorage "github.com/openshift/origin/pkg/oc/cli/admin/migrate/storage"
-	migratetemplateinstances "github.com/openshift/origin/pkg/oc/cli/admin/migrate/templateinstances"
 	"github.com/openshift/origin/pkg/oc/cli/admin/mustgather"
-	"github.com/openshift/origin/pkg/oc/cli/admin/network"
-	"github.com/openshift/origin/pkg/oc/cli/admin/node"
 	"github.com/openshift/origin/pkg/oc/cli/admin/policy"
 	"github.com/openshift/origin/pkg/oc/cli/admin/project"
 	"github.com/openshift/origin/pkg/oc/cli/admin/prune"
 	"github.com/openshift/origin/pkg/oc/cli/admin/release"
 	"github.com/openshift/origin/pkg/oc/cli/admin/top"
-	"github.com/openshift/origin/pkg/oc/cli/admin/upgrade"
 	"github.com/openshift/origin/pkg/oc/cli/admin/verifyimagesignature"
 	"github.com/openshift/origin/pkg/oc/cli/kubectlwrappers"
 )
