+++ pkg/oc/cli/admin/groups/groups.go
@@ -6,9 +6,9 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
-	"github.com/openshift/origin/pkg/oc/cli/admin/groups/new"
+	"github.com/openshift/oc/pkg/cli/admin/groups/new"
+	"github.com/openshift/oc/pkg/cli/admin/groups/users"
 	"github.com/openshift/origin/pkg/oc/cli/admin/groups/sync"
-	"github.com/openshift/origin/pkg/oc/cli/admin/groups/users"
 )
 
 const GroupsRecommendedName = "groups"
