+++ staging/src/github.com/openshift/oc/pkg/cli/admin/cert/pem.go
