+++ staging/src/github.com/openshift/oc/pkg/cli/admin/createerrortemplate/create_error_template.go
