+++ staging/src/github.com/openshift/oc/pkg/cli/admin/createproviderselectiontemplate/create_provider_selection_template.go
