+++ staging/src/github.com/openshift/oc/pkg/cli/admin/groups/users/remove.go
