+++ staging/src/github.com/openshift/oc/pkg/cli/admin/migrate/legacyhpa/hpa.go
@@ -16,7 +16,7 @@ import (
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
-	"github.com/openshift/origin/pkg/oc/cli/admin/migrate"
+	"github.com/openshift/oc/pkg/cli/admin/migrate"
 )
 
 var (
