+++ staging/src/github.com/openshift/oc/pkg/cli/admin/migrate/legacyhpa/hpa_test.go
