+++ staging/src/github.com/openshift/oc/pkg/cli/admin/migrate/migrator_test.go
