+++ staging/src/github.com/openshift/oc/pkg/cli/admin/migrate/templateinstances/templateinstances_test.go
