+++ staging/src/github.com/openshift/oc/pkg/cli/admin/network/isolate_projects.go
