+++ staging/src/github.com/openshift/oc/pkg/cli/admin/network/join_projects.go
