+++ staging/src/github.com/openshift/oc/pkg/cli/admin/network/make_projects_global.go
