+++ staging/src/github.com/openshift/oc/pkg/cli/admin/network/project_options.go
