+++ staging/src/github.com/openshift/oc/pkg/cli/admin/node/logs_test.go
