+++ staging/src/github.com/openshift/oc/pkg/cli/admin/prune/builds/prune.go
