+++ staging/src/github.com/openshift/oc/pkg/cli/admin/prune/deployments/resolvers_test.go
