+++ test/integration/oauth_ldap_test.go
@@ -23,9 +23,9 @@ import (
 	restclient "k8s.io/client-go/rest"
 
 	authapi "github.com/openshift/oauth-server/pkg/api"
+	"github.com/openshift/oc/pkg/cli/admin/cert"
 	"github.com/openshift/oc/pkg/helpers/tokencmd"
 	configapi "github.com/openshift/origin/pkg/cmd/server/apis/config"
-	"github.com/openshift/origin/pkg/oc/cli/admin/cert"
 	userclient "github.com/openshift/origin/pkg/user/generated/internalclientset/typed/user/internalversion"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
